+++ scieloapi/__init__.py
@@ -1,11 +1,5 @@
-try:
-    from scieloapi import Connector, Endpoint, Client
-except ImportError as e:
-    if 'requests' in e.message:
-        print 'requests is not present. please run python setup.py install'
-    else:
-        raise
-
-
 __version__ = '0.4'
+__user_agent__ = 'scieloapi/%s' % __version__
+
+from .core import Connector, Endpoint, Client